Restructure analysis
