--- e2e_sae/scripts/dashboards.yaml
@@ -1,22 +1,0 @@
-pretrained_sae_paths: null # Paths of the pretrained SAEs to load. Should be a path to a .pt file, or a list of them. Can also be provided as a second argument in the command line.
-sae_config_path: null # Path to the config file used to train the SAEs (if null, we'll assume it's at pretrained_sae_paths[0].parent / "config.yaml")
-n_samples: 10_000
-batch_size: 64
-minibatch_size_features: 128 # Num features in each batch of calculations. Lower to avoid OOM errors
-data: # DatasetConfig for the data which will be used to generate the dashboards
-  dataset_name: 'apollo-research/roneneldan-TinyStories-tokenizer-gpt2'
-  is_tokenized: True
-  tokenizer_name: 'gpt2'
-  split: "validation"
-  n_ctx: 512
-save_dir: null # The directory for saving the HTML feature dashboard files
-sae_positions: null # The names of the SAE positions to generate dashboards for. e.g.'blocks.2.hook_resid_post'. If None, then all positions will be generated
-feature_indices: null # The features for which to generate dashboards on each SAE. If none, then we'll generate dashbaords for every feature.
-prompt_centric: # Used to generate prompt-centric (rather than feature-centric) dashboards. Feature-centric dashboards will also be generated for every feature appaearing in these
-  n_random_prompt_dashboards: 50 # The number of random prompts to generate prompt-centric dashboards for.
-  data: null # "DatasetConfig for getting random prompts. If None, then non-prompt-centric data will be used
-  prompts: # Specific prompts on which to generate prompt-centric feature dashboards. A feature-centric dashboard will be generated for every token position in each prompt.
-    - "Sally met Mike at the show. She brought popcorn for him."
-  str_score: "loss_effect" # The ordering metric for which features are most important in prompt-centric dashboards. Can be one of 'act_size', 'act_quantile', or 'loss_effect'
-  num_top_features: 10 # How many of the most relevant features to show for each prompt in the prompt-centric dashboards
-seed: 0
