+++ pyproject.toml
@@ -75,7 +75,7 @@ known-third-party = ["wandb"]
 [tool.pyright]
 include = ["e2e_sae", "tests"]
 exclude = [
-    "e2e_sae/scripts/autointerp.py",  # neuron_explainer package not yet on pypi
+    "e2e_sae/scripts/analysis/autointerp.py",  # neuron_explainer package not yet on pypi
 ]
 
 strictListInference = true
